update authcontroller.php, besercontroller.php & api.php

+++ app/Http/Controllers/BeserController.php
@@ -28,15 +28,33 @@ class BeserController extends Controller
     
     public function show($id)
     {
+        $beser = Beser::find($id);
+        if(empty($beser)){
+            return response()->json(['error' => 'Id '.$id.' not found'], 401);
+        }
+        
         return new BeserResource(
-            Beser::find($id)
+            $beser
         );
     }
 
-    public function search($lang, $word) 
+    public function search(Request $request) 
     {
+        $validator = Validator::make(request()->all(), [
+            'language' => 'required',
+            'word' => 'required',
+        ]);
+
+        if($validator->fails()) {
+            response()->json($validator->messages(), 422)->send();
+            exit;
+        }
+        
+        $lang = request('language');
+        $word = request('word');
+
         if($lang != "indonesia" && $lang != "beser"){
-            return response()->json(['warning' => 'Parameter ('.$lang.') Not Valid. Recomended (indonesia or beser)'], 422);
+            return response()->json(['orror' => 'Parameter ('.$lang.') Not Valid. Recomended (indonesia or beser)'], 401);
         }
 
         return BeserResource::collection(
@@ -60,16 +78,18 @@ class BeserController extends Controller
     public function update(Request $request, $id)
     {
         $user = $this->getAuthUser();
-
         $this->validateRequest();
-        $beser = Beser::find($id);
-        $this->checkOwnership($beser->user_id);
 
+        $beser = Beser::find($id);
+        if(empty($beser)){
+            return response()->json(['error' => 'Id '.$id.' not found'], 401);
+        }
+
+        $this->checkOwnership($beser->user_id);
         $beser->update([
             'indonesia' => request('indonesia'), 
             'beser' => request('beser')
         ]);
-        
         return response()->json(['message' => 'Successfully updated']);
     }
     
@@ -78,8 +98,11 @@ class BeserController extends Controller
         $this->getAuthUser();
 
         $beser = Beser::find($id);
+        if(empty($beser)){
+            return response()->json(['error' => 'Id '.$id.' not found'], 401);
+        }
+        
         $this->checkOwnership($beser->user_id);
-
         $beser->delete();      
 
         return response()->json(['message' => 'Successfully deleted']);
@@ -97,6 +120,4 @@ class BeserController extends Controller
             exit;
         }
     }
-
-    
 }
